refactor(skill): consolidate skill modules under openkb/skill/
All skill-related code now lives in a single openkb/skill/ subpackage
(7 modules + __init__). Drop the redundant 'skill_' prefix on filenames
since the package qualifies them already.

Moves:
  openkb/generator.py            -> openkb/skill/generator.py
  openkb/marketplace.py          -> openkb/skill/marketplace.py
  openkb/skill_validator.py      -> openkb/skill/validator.py
  openkb/skill_workspace.py      -> openkb/skill/workspace.py
  openkb/agent/skill_creator.py  -> openkb/skill/creator.py
  openkb/agent/skill_tools.py    -> openkb/skill/tools.py
  openkb/agent/skill_evaluator.py -> openkb/skill/evaluator.py

generator.py + marketplace.py go under skill/ for now (v0.x only has
skills); they're nominally generic primitives but YAGNI -- when a
second artifact type (ppt / podcast / report) actually lands, those
two will move back out to openkb/<shared>/.

No behavioral changes. All imports + test patch targets updated.
Test suite stays at 494 passed.
