+++ tests/test_skill_validator.py
@@ -1,10 +1,10 @@
-"""Unit tests for openkb.skill_validator — pure-Python structural checks
+"""Unit tests for openkb.skill.validator — pure-Python structural checks
 on a compiled skill directory. No LLM, no network."""
 from __future__ import annotations
 
 from pathlib import Path
 
-from openkb.skill_validator import (
+from openkb.skill.validator import (
     DESCRIPTION_MAX_CHARS,
     REFERENCE_MAX_BYTES,
     SKILL_MD_MAX_BYTES,
