+++ tests/test_skill_workspace.py
@@ -1,11 +1,11 @@
-"""Tests for :mod:`openkb.skill_workspace` — iteration save/restore + diff."""
+"""Tests for :mod:`openkb.skill.workspace` — iteration save/restore + diff."""
 from __future__ import annotations
 
 from pathlib import Path
 
 import pytest
 
-from openkb.skill_workspace import (
+from openkb.skill.workspace import (
     list_iterations,
     restore_iteration,
     save_iteration,
